requirements.txt
# Core dependencies
GitPython>=3.1.40
openai>=1.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0